renamed logging.py to my_logging.py, and github to my_github
