+++ src/ogd_to_lod/metadata/generator.py
@@ -18,7 +18,7 @@ import re
 from typing import TYPE_CHECKING, Any
 
 from ogd_to_lod._slug import slugify
-from ogd_to_lod.logging import get_logger
+from ogd_to_lod.my_logging import get_logger
 
 if TYPE_CHECKING:
     from ogd_to_lod.lookup import ReuseContext
