+++ src/ogd_to_lod/rml/generator.py
@@ -6,7 +6,7 @@ from typing import Any
 from ogd_to_lod._slug import slugify
 from ogd_to_lod.ai import AIService
 from ogd_to_lod.lookup import ReuseContext
-from ogd_to_lod.logging import get_logger
+from ogd_to_lod.my_logging import get_logger
 from ogd_to_lod.rml.prompts import RML_CORRECTION_PROMPT, RML_GENERATION_PROMPT
 
 logger = get_logger(__name__)
